Fixed: Elasticsearch5 retry logic to sleep per request failure instead of batch of failures
Before this change ES5 could get into a state where large batches would fail again and again
because the retry logic would try a batch, sleep, then try the whole batch again, which would
typically fail. This change makes it so that a batch is tried, if any of those requests in the
batch fails the time between requests will be increased until the requests no longer fail.