Invalidate the define property cache
When we get a metadata change event that is from a define property
change, also invalidate the cache in GraphBase.

Fixes issues where multiple graph processes can start seeing elements
and properties that don't exist in the graphs local caches.

Was found because publish/undo needs the properties to be defined and
those run in the worker graph thread.